Refactor: change file structure
