+++ client/src/components/Footer.jsx
@@ -3,20 +3,6 @@ import PropTypes from 'prop-types';
 import Box from '@mui/material/Box';
 import Container from '@mui/material/Container';
 import Typography from '@mui/material/Typography';
-import Link from '@mui/material/Link';
-
-function Copyright() {
-  return (
-    <Typography variant="body2" color="text.secondary" align="center">
-      {'Copyright © '}
-      <Link color="inherit" href="https://mui.com/">
-        Your Website
-      </Link>{' '}
-      {new Date().getFullYear()}
-      {'.'}
-    </Typography>
-  );
-}
 
 function Footer(props) {
   const { description, title } = props;
@@ -35,7 +21,6 @@ function Footer(props) {
         >
           {description}
         </Typography>
-        <Copyright />
       </Container>
     </Box>
   );
