+++ client/src/components/MainFeaturedPost.jsx
