+++ client/src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 //import App from './App.jsx'
-import Blog from './blog/Blog.jsx'
+import Blog from './Blog'
 import './main.css';
